@@ -1,12 +1,14 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
+    ext.kotlin_version = '1.4.30'
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.0.1'
+        classpath 'com.android.tools.build:gradle:4.1.2'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -32,15 +34,15 @@ task clean(type: Delete) {
 ext {
     // Sdk and tools
     // Support library and architecture components support minSdk 14 and above.
-    buildToolsVersion = '26.0.2'
-    compileSdkVersion = 26
-    minSdkVersion = 14
-    targetSdkVersion = 26
+    buildToolsVersion = '30.0.2'
+    compileSdkVersion = 30
+    minSdkVersion = 16
+    targetSdkVersion = 30
 
     // App dependencies
     supportLibraryVersion = '26.+'
     junitVersion = '4.12'
-    butterknifeVersion = '8.8.1'
+    butterknifeVersion = '10.2.0'
     okhttpVersion = '3.8.1'
     retrofitVersion = '2.3.0'
     rxjavaVersion = '1.3.0'
@@ -48,5 +50,5 @@ ext {
     gsonVersion = '2.8.1'
     glideVersion = '3.7.0'
     bannerVersion = '1.4.10'
-    daggerVersion = '2.14.1'
+    daggerVersion = '2.24'
 }
