@@ -1,24 +1,23 @@
 package com.android.march.one.adapter.holder;
 
 import android.content.res.Resources;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.LinearLayout;
 import android.widget.RatingBar;
 import android.widget.TextView;
 
-import com.android.march.one.OneUtils;
 import com.android.march.one.R;
 import com.android.march.one.adapter.RecyclerViewAdapter;
-import com.android.march.one.model.bean.MovieListBean;
+import com.android.march.one.model.bean.MovieBean;
 import com.android.march.one.ui.activity.MovieDetailActivity;
 import com.bumptech.glide.Glide;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 
-public class MovieListViewHolder extends RecyclerViewAdapter.ViewHolder<MovieListBean.MovieBean> {
+public class MovieListViewHolder extends RecyclerViewAdapter.ViewHolder<MovieBean> {
 
     private final AppCompatActivity activity;
     private final int type;
@@ -52,22 +51,22 @@ public class MovieListViewHolder extends RecyclerViewAdapter.ViewHolder<MovieLis
     }
 
     @Override
-    protected void bindViewHolder(RecyclerViewAdapter<MovieListBean.MovieBean> adapter, int position, MovieListBean.MovieBean movieBean) {
-        Glide.with(activity).load(movieBean.getImageBean().getLarge()).placeholder(R.drawable.ic_movie_placeholder).into(ivMovie);
+    protected void bindViewHolder(RecyclerViewAdapter<MovieBean> adapter, int position, MovieBean movieBean) {
+        Glide.with(activity).load(movieBean.getImg()).placeholder(R.drawable.ic_movie_placeholder).into(ivMovie);
 
         tvMovieTitle.setText(movieBean.getTitle());
-        tvMovieDirector.setText(resources.getString(R.string.movie_director, OneUtils.formatDirectorName(movieBean.getDirectorBeanList())));
-        tvMovieCast.setText(resources.getString(R.string.movie_star, OneUtils.formatCastName(movieBean.getCastBeanList())));
-        tvMovieGenre.setText(resources.getString(R.string.movie_classify, OneUtils.formatGenre(movieBean.getGenreList())));
+        tvMovieDirector.setText(resources.getString(R.string.movie_director, movieBean.getDN()));
+        tvMovieCast.setText(resources.getString(R.string.movie_star, movieBean.getActors()));
+        tvMovieGenre.setText(resources.getString(R.string.movie_classify, movieBean.getMovieType()));
 
         if (type == 0) {
             llRating.setVisibility(View.VISIBLE);
-            ratingBar.setRating(movieBean.getRatingBean().getAverage() / 2);
-            tvMovieRating.setText(String.valueOf(movieBean.getRatingBean().getAverage()));
+            ratingBar.setRating(movieBean.getR() / 2);
+            tvMovieRating.setText(String.valueOf(movieBean.getR()));
         } else {
             llRating.setVisibility(View.GONE);
         }
 
-        llItem.setOnClickListener(v -> MovieDetailActivity.start(activity, movieBean.getId(), movieBean.getTitle(), movieBean.getImageBean().getLarge(), ivMovie));
+        llItem.setOnClickListener(v -> MovieDetailActivity.start(activity, movieBean.getId() + "", movieBean.getTCn(), movieBean.getImg(), ivMovie));
     }
 }
