@@ -1,6 +1,6 @@
 package com.android.march.one.adapter.holder;
 
-import android.support.v7.widget.CardView;
+import androidx.cardview.widget.CardView;
 import android.view.View;
 import android.widget.ImageView;
 import android.widget.TextView;
