@@ -3,7 +3,7 @@ package com.android.march.one.ui.activity;
 import android.content.Context;
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 import com.android.march.one.StatusBarUtils;
 import com.android.march.one.base.BaseActivity;
