@@ -2,8 +2,8 @@ package com.android.march.one.ui.fragment;
 
 import android.annotation.SuppressLint;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.Menu;
 import android.view.MenuInflater;
@@ -13,7 +13,6 @@ import android.view.ViewGroup;
 
 import com.android.march.one.LazyLoadFragment;
 import com.android.march.one.R;
-import com.android.march.one.ui.activity.SearchActivity;
 import com.android.march.one.ui.view.MusicView;
 
 import java.util.List;
