@@ -1,8 +1,8 @@
 package com.android.march.one.ui.view;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
+import androidx.recyclerview.widget.RecyclerView;
+import androidx.appcompat.widget.Toolbar;
 import android.util.AttributeSet;
 
 import com.android.march.one.Constants;
@@ -81,7 +81,7 @@ public class HomeView extends RootView<HomeContract.Presenter> implements HomeCo
 
     // 没请求到数据就一直请求前一天数据
     public void getData() {
-        presenter.getDay(year, month, day);
+        // presenter.getDay(year, month, day);
     }
 
     @Override
