@@ -1,8 +1,8 @@
 package com.android.march.one.ui.view;
 
 import android.content.Context;
-import android.support.v7.widget.SwitchCompat;
-import android.support.v7.widget.Toolbar;
+import androidx.appcompat.widget.SwitchCompat;
+import androidx.appcompat.widget.Toolbar;
 import android.util.AttributeSet;
 import android.widget.RelativeLayout;
 import android.widget.TextView;
